+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [UNRELEASED]
 
-This version is based on Valgrind-3.17.0.
+This version is based on Valgrind-3.19.0.
 
 ### Added