+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## [UNRELEASED]
+## v2.4.0 - 2022-12-15
 
 This version is based on Valgrind-3.20.0.